+++ include/proxsuite-nlp/prox-solver.hpp
@@ -49,8 +49,11 @@ public:
   using ConstraintSet = ConstraintSetBase<Scalar>;
   using ConstraintObject = ConstraintObjectTpl<Scalar>;
 
-  /// Manifold on which to optimize.
-  shared_ptr<Problem> problem_;
+protected:
+  /// General nonlinear program to solve.
+  Problem *problem_;
+
+public:
   /// Merit function.
   ALMeritFunctionTpl<Scalar> merit_fun;
   /// Proximal regularization penalty.
@@ -102,18 +105,16 @@ public:
 
   //// Parameters for the inertia-correcting strategy.
 
-  const Scalar del_inc_k = 8.;
-  const Scalar del_inc_big = 100.;
-  const Scalar del_dec_k = 1. / 3.;
-
-  const Scalar DELTA_MIN = 1e-14; // Minimum nonzero regularization strength.
-  const Scalar DELTA_MAX = 1e6;   // Maximum regularization strength.
-  const Scalar DELTA_NONZERO_INIT = 1e-4;
+  Scalar del_inc_k = 8.;      //< Inertia corrector increase factor
+  Scalar del_inc_big = 100.;  //< Inertia corrector increase factor (big)
+  Scalar del_dec_k = 1. / 3.; //< Inertia corrector decrease factor
+  Scalar DELTA_MIN = 1e-14;   //< Minimum nonzero regularization strength.
+  Scalar DELTA_MAX = 1e6;     //< Maximum regularization strength.
+  Scalar DELTA_NONZERO_INIT = 1e-4;
   Scalar DELTA_INIT = 0.;
 
-  /// Solver maximum number of iterations.
-  std::size_t max_iters = 100;
-  std::size_t max_al_iters = 1000;
+  std::size_t max_iters = 100;     //< Solver maximum number of iterations.
+  std::size_t max_al_iters = 1000; //< Maximum outer (AL) iterations.
 
   /// Callbacks
   std::vector<CallbackPtr> callbacks_;
@@ -121,7 +122,7 @@ public:
   unique_ptr<Workspace> workspace_;
   unique_ptr<Results> results_;
 
-  ProxNLPSolverTpl(shared_ptr<Problem> prob, const Scalar tol = 1e-6,
+  ProxNLPSolverTpl(Problem &prob, const Scalar tol = 1e-6,
                    const Scalar mu_eq_init = 1e-2, const Scalar rho_init = 0.,
                    const VerboseLevel verbose = QUIET,
                    const Scalar mu_lower = 1e-9, const Scalar prim_alpha = 0.1,
@@ -129,6 +130,8 @@ public:
                    const Scalar dual_beta = 1.,
                    LDLTChoice ldlt_blocked = LDLTChoice::BUNCHKAUFMAN,
                    const LinesearchOptions ls_options = LinesearchOptions());
+
+  const Problem &problem() const { return *problem_; }
 
   const Manifold &manifold() const { return *problem_->manifold_; }